Retail-project/csv_uploader/requirements.txt
flask
boto3
python-dotenv 